passive avoidance - last/requirements.txt
altgraph==0.17.4
asgiref==3.7.2
asttokens==2.4.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.9
Django==4.2.7
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.43.1
fsspec==2024.6.1
future==1.0.0
idna==3.7
image==1.5.33
imageio==2.33.0
intel-openmp==2021.4.0
ipython==8.25.0
iso8601==2.1.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lapx==0.5.9.post1
lazy_loader==0.3
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.1.7
minimalmodbus==2.1.1
mkl==2021.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.1
opencv-python==4.11.0.86
opencv-python-headless==4.8.1.78
openpyxl==3.1.4
packaging==23.2
pandas==2.2.2
parso==0.8.4
pefile==2023.2.7
Pillow==10.0.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.18.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQt5Designer==5.14.1
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
pyserial==3.5
PySide6==6.8.1
PySide6_Addons==6.8.1
PySide6_Essentials==6.8.1
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.2
serial==0.0.97
setuptools==70.0.0
shapely==2.0.5
shiboken6==6.8.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.13.0
tbb==2021.13.0
thop==0.1.1.post2209072238
tifffile==2023.9.26
tk==0.1.0
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
ttkbootstrap==1.10.1
typing_extensions==4.12.2
tzdata==2023.3
ultralytics==8.3.140
ultralytics-thop==2.0.0
urllib3==2.2.2
virtualenv==20.28.0
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.45.1
